+++ matdb/database/controller.py
@@ -6,16 +6,204 @@ from matdb import msg
 import numpy as np
 import six
 
-class Sequence(object):
-    """Represents a sequence of groups that form a database (all inheriting from 
-    :class:`Group`) that are all related be the atomic configuration that they model.
+import collections
+
+def db_pgrid(options, ignore_=None):
+    """Creates a parameter grid over the specified options for the database.
+
+    .. note:: This function treats keys that end in `*` specially. When the key
+      ends in `*`, the values in the specified list contribute to the cartesian
+      product.
+
+    Args:
+        options (dict): key-value pairs to iterate across.
+        ignore_ (list): of `str` keys to ignore in the options dict.
+
+    Returns:
+        tuple: `(grid, keys, suffix_grid)` where grid is a list of 
+        tuples with a value for each key in `options`, keys is a liste of 
+        the keys in the order they appear in the tuples, suffix_grid is the 
+        list of the suffix compinations for the grids.
+    """
+    from itertools import product
+    from collections import OrderedDict
+    from operator import itemgetter
+    from matdb.utility import special_values
+
+    ignore = ["root","atoms","parent"]
+    if ignore_ is not None:
+        for i in ignore_:
+            ignore.append(i)
+    #Sort the dict so that we get the same ordering of parameters in the
+    #product.
+    params = OrderedDict(sorted(options.items(), key=itemgetter(0)))
+    values = []
+    keys = []
+    suffix_keys = []
+    suffix_values = []
+    for k, v in params.items():
+        if (ignore is not None and k in ignore or "_suffix" in k) or "*" not in k:
+            continue
+
+        values.append(v)
+        keys.append(k.strip('*'))
+        suff_k = "{0}_suffix*".format(k.strip('*'))
+        if suff_k in params:
+            suffix_values.append([float(i) if isinstance(i,(int,float)) else i for i in params[suff_k]])
+        elif suff_k.strip('*') in params:
+            try:
+                from matdb.utility import special_functions
+                suffix_values.append([float(i) for i in special_functions(params[suff_k.strip('*')],v)])
+            except:
+                msg.warn("The {} function is not supported. Using the default suffix.".format(params[suff_k.strip('*')]))
+                suffix_values.append([float(i) for i in range(1,len(v)+1)])
+        else:
+            suffix_values.append([float(i) for i in range(1,len(v)+1)])
+        suffix_keys.append(suff_k.strip('*'))
+            
+    suffix_grid = list(product(*suffix_values))
+    grid = list(product(*values))
+    
+    if len(grid) != len(suffix_grid): #pragma: no cover
+        msg.err("The grid and the suffix grid don't match.")
+        
+    return (grid, keys, suffix_grid)
+
+class ParameterGrid(collections.MutableSet):
+    """An ordered list of the paramater combinations for the database. 
+    Values are the suffixes of the combinations of parameters as tuples:
+    e.g. (8, "dog", 1.2) for "dim", "animal", "temperature"
+    ({"animal*": ["dog", "cat", "cow"], "dim*": [[],[],[]], "temperature": 1.2})
+
+    Args:
+        params (dict): the paramaters needed to build the database.
+    
+    Attributes:
+        values (dict): keys are the suffix tuple and the values are the 
+            actual values needed by the database.
+        keys (list): the `str` names of the different parameters in the database.
+    """
+    def __init__(self, params):
+        from matdb.utility import flatten_dict
+        (items,names,ids) = db_pgrid(flatten_dict(params))
+        #add these items to the set.
+        self.end = end = [] 
+        end += [None, end, end]         # sentinel node for doubly linked list
+        self.map = {}                   # key --> [key, prev, next]
+        self.keys = names 
+        self.values = {}
+        for i, v in zip(ids,items):
+            self.add(i,v)
+
+    def to_str(self, key):
+        """Converts the tuple into a folder name with the 
+        parameter name and value truncated and converted to 
+        the correct format.
+        
+        Args:
+            key (tuple): the key to be converted.
+            
+        Returns:
+            folder (str): the name of the folder.
+        """
+        from matdb.utility import is_number
+        
+        temp = []
+        for n,i in zip(self.keys,key):
+            if is_number(i) and i.is_integer():
+                t_i = str(int(i))
+            else:
+                t_i = str(i)
+            temp.append("-".join([n[:4],t_i]))
+        folder = "-".join(temp)
+        
+        return folder
+        
+    def from_str(self, folder):
+        """Converts the folder name back to it's original format.
+        
+        Args:
+            folder (str): the folder name to be converted.
+            
+        Returns:
+            key (tuple): the tuple the folder was made from.
+        """
+        from matdb.utility import is_number
+        temp = folder.strip().split("-")
+        key = tuple([float(i) if is_number(i) else i for i in temp[1::2]])
+        return key        
+            
+    def __len__(self):
+        return len(self.map)
+
+    def __contains__(self, key):
+        return key in self.map
+
+    def __getitem__(self,key):
+        return self.values[key]
+
+    def add(self, key,value):
+        """Adds key to the set if it is not already in the set.
+
+        Args:
+            key (tuple): Anything that could be added to the set.
+            value (tuple): The actual values that the suffix's 
+                correspond to.
+        """
+        if key not in self.map:
+            end = self.end
+            curr = end[1]
+            curr[2] = end[1] = self.map[key] = [key, curr, end]
+            self.values[key] = value
+        else:
+            msg.warn("The key {} already exists in the set, ignoring addition.".format(key))
+
+    def discard(self, key):
+        """Removes the key from the set.
+
+        Args:
+            key (tuple): An element of the set.
+        """        
+        if key in self.map:        
+            key, prev, next = self.map.pop(key)
+            prev[2] = next
+            next[1] = prev
+            self.values.pop(key,None)
+
+    def __iter__(self):
+        end = self.end
+        curr = end[2]
+        while curr is not end:
+            yield curr[0]
+            curr = curr[2]
+
+    def pop(self, key):
+        """Removes an element from the set.
+
+        Args:
+            key (tuple): An element of the set.
+        """
+        self.discard(key)
+        return key
+
+    def __repr__(self):
+        if not self:
+            return '%s()' % (self.__class__.__name__,)
+        return '%s(%r)' % (self.__class__.__name__, list(self))
+
+    def __eq__(self, other):
+        if isinstance(other, ParameterGrid):
+            return len(self) == len(other) and list(self) == list(other)
+        return set(self) == set(other)
+
+class Database(object):
+    """Represents a Database of groups (all inheriting from :class:`Group`) that 
+    are all related be the atomic configuration that they model.
     .. note:: See the list of attributes below which correspond to the sections
       in the YAML database specification file.
     Args:
         name (str): name of the configuration that this database sequence is
           operating for.
-        repeater (Repeater): repeater for a set of sequences that all
-          share the same root `atoms` object.
         root (str): root directory in which all other database sequences for
           the configurations in the same specification will be stored.
         parent (Controller): instance controlling multiple configurations.
@@ -43,12 +231,12 @@ class Sequence(object):
           etc.); values are the corresponding class instances.
         parent (Controller): instance controlling multiple configurations.
     """
-    def __init__(self, name, repeater, root, parent, steps, splits):
+    def __init__(self, name, root, parent, steps, splits):
+        import pudb
+        pudb.set_trace()
         self.name = name
         self.config = name.split('.')[0]
-        self.atoms = repeater.atoms.copy()
-        self.root = path.join(root, name)
-        self.repeater = repeater
+        self.root = root
         self.splits = {} if splits is None else splits
         
         if not path.isdir(self.root):
@@ -90,7 +278,8 @@ class Sequence(object):
             #pointers; then add in the keyword arguments that are missing.
             cpspec = dbspec.copy()
             del cpspec["type"]
-            cpspec["atoms"] = self.atoms
+            cpspec["atoms"] = dbspec["seed"]
+            del cpspec["seed"]
             cpspec["root"] = self.root
             cpspec["parent"] = self
 
@@ -99,10 +288,21 @@ class Sequence(object):
             for k in list(cpspec.keys()):
                 if isinstance(cpspec[k], dict) and self.config in cpspec[k]:
                     cpspec[k] = cpspec[k][self.config]
+
+            for k in dbspec.keys():
+                if k[-1] == '*':
+                    del cpspec[k]
+                    cpspec[k[0:-1]] = ParameterGrid(dbspec[k])
             
+            import pudb
+            pudb.set_trace()
             instance = cls(**cpspec)
             self.steps[instance.name] = instance
 
+        ## Fix this latter
+        if self.steps[0] is not seeded:
+            raise ValueError()
+            
     @property
     def isteps(self):
         """Returns a generator over steps in this sequence. The generator yields
@@ -321,157 +521,7 @@ class Sequence(object):
             msg.info("Setting up database {}:{}".format(self.name, dbname))
             db.setup(rerun)
         msg.blank()
-
-class Repeater(object):
-    """Repeats sequences of database steps across a parameter grid.
-    Args:
-        name (str): name of the configuration that this database sequence is
-          operating for.
-        poscar (str): name of the POSCAR file in `root` to extract atomic
-          configuration information from.
-        root (str): root directory in which all other database sequences for
-          the configurations in the same specification will be stored.
-        parent (Controller): instance controlling multiple configurations.
-        steps (list): of `dict` describing the kinds of sub-configuration
-          database steps to setup.
-        splits (dict): keys are split names; values are `float` *training*
-          percentages to use.
-    Attributes:
-        atoms (quippy.atoms.Atoms): a single atomic configuration from
-          which many others may be derived using MD, phonon
-          displacements, etc.
-        poscar (str): path to the POSCAR file for the seed configuration that
-          all sequences in this repeater will use.
-        kfile (str): path to the kpath.json cached file for this repeater.        
-    """
-    def __init__(self, name, poscar, root, parent, steps, niterations=None,
-                 splits=None):
-        from collections import OrderedDict
-        from copy import copy
-        from quippy.atoms import Atoms
         
-        self.name = name
-        self.sequences = OrderedDict()
-        self.poscar = path.join(root, poscar)
-        self.atoms = Atoms(self.poscar, format="POSCAR")
-        self.kfile = path.join(parent.kpathdir, "{0}.json".format(self.name))
-        self.steps = steps
-
-        self._kpath = None
-        """tuple: result of querying the materialscloud.org special path
-        service. First term is a list of special point labels; second is the
-        list of points corresponding to those labels.
-        """
-        
-        if niterations is not None:
-            from matdb.utility import obj_update
-            for i, repeater in enumerate(niterations):
-                nname = None
-                isteps = copy(steps)
-                for k, v in repeater.items():
-                    if k == "suffix":
-                        nname = self.name + v
-                        continue
-                    
-                    obj_update(isteps, k, v, False)
-                    
-                if nname is None:
-                    nname = self.name + "-{0:d}".format(i)
-
-                iobj = Sequence(nname, self, root, parent, isteps, splits)
-                self.sequences[nname] = iobj
-        else:
-            single = Sequence(name, self, root, parent, steps, splits)
-            self.sequences[name] = single
-
-    @property
-    def kpath(self):
-        """Returns the materialscloud.org special path in k-space for the seed
-        configuration of this database.
-        Returns:
-            tuple: result of querying the materialscloud.org special path
-            service. First term is a list of special point labels; second is the
-            list of points corresponding to those labels.
-        """
-        if self._kpath is None:
-            import json
-            #We use some caching here so that we don't have to keep querying the
-            #server and waiting for an identical response.
-            if path.isfile(self.kfile):
-                with open(self.kfile) as f:
-                    kdict = json.load(f)
-            else:
-                from .phonon import _parsed_kpath
-                labels, band = _parsed_kpath(self.poscar)
-                kdict = {"labels": labels, "band": band}
-                with open(self.kfile, 'w') as f:
-                    json.dump(kdict, f)
-            
-            self._kpath = (kdict["labels"], kdict["band"])
-            
-        return self._kpath
-            
-    def recover(self, rerun=False):
-        """Runs recovery on each step in the sequence to determine which configs failed
-        and then create a jobfile to requeue them for compute.
-        Args:
-            rerun (bool): when True, recreate the jobfile even if it
-              already exists.
-        """
-        for db in self.sequences.values():
-            db.recover(rerun)
-
-    def status(self, busy=False):
-        """Prints a status message for each step in the sequence relative
-        to VASP execution status.
-        Args:
-            busy (bool): when True, print a list of the configurations that are
-              still busy being computed in DFT.
-        """
-        for db in self.sequences.values():
-            db.status(busy)
-
-    def execute(self, recovery=False, env_vars=None):
-        """Submits job array files for any of steps in the sequence that are ready to
-        execute, but which haven't been submitted yet.
-        Args:
-            recovery (bool): when True, submit the script for running recovery
-              jobs.
-            env_vars (dict): of environment variables to set before calling the
-              execution. The variables will be set back after execution.
-        """
-        for db in self.sequences.values():
-            db.execute(recovery, env_vars=env_vars)
-
-    def split(self, recalc=0):
-        """Splits the total available data in each step's databases into a training and
-        holdout set.
-        Args:
-            recalc (int): when non-zero, re-split the data and overwrite any
-              existing *.xyz files. This parameter decreases as
-              rewrites proceed down the stack. To re-calculate
-              lower-level XYZ files, increase this value.
-        """
-        for db in self.sequences.values():
-            nrecalc = recalc - (0 if len(self.steps) == 0 else 1)
-            db.split(nrecalc)
-
-    def cleanup(self):
-        """Runs the cleanup methods of each step's databases in the sequence.
-        """
-        for db in self.sequences.values():
-            db.cleanup()
-
-    def setup(self, rerun=False):
-        """Sets up the each step's database in the sequence in order. If an
-        earlier step is not ready yet, the next item in the step won't setup.
-        Args:
-            rerun (bool): when True, recreate the folders even if they
-              already exist. 
-        """
-        for db in self.sequences.values():
-            db.setup(rerun)
-            
 class Controller(object):
     """Implements methods for tying a configuration dictionary (in
     YAML format) to instances of various databases.
@@ -495,6 +545,8 @@ class Controller(object):
           potentials after fitting.
     """
     def __init__(self, config, tmpdir=None):
+        import pudb
+        pudb.set_trace()
         from matdb.io import read
         self.config = path.expanduser(path.abspath(config))
         if path.isabs(config):
@@ -512,8 +564,8 @@ class Controller(object):
         self.plotdir = path.join(self.root, "plots")
         self.kpathdir = path.join(self.root, "kpaths")
         self.title = self.specs["title"]
-        self.seeded = {}
         self.seedless = {}
+        self.seeded = {}
         self.species = [s for s in self.specs["species"]]
         self.potcars = self.specs.get("potcars", {})
         self.incar = self.specs.get("incar", {})
@@ -525,7 +577,9 @@ class Controller(object):
         # the sequences.
         from matdb.database.legacy import LegacyDatabase
         for dbspec in self.specs.get("databases", []):
-            if dbspec.get("legacy", False):
+            seedless_types = ["legacy","enumerated"]
+            test = [i in dbspec for i in seedless_types]
+            if any(test):
                 cpspec = dbspec.copy()
                 cpspec["root"] = self.root
                 cpspec["controller"] = self
@@ -533,8 +587,10 @@ class Controller(object):
                 #We allow the user to specify the folder relative to repository
                 #root; this is mainly for unit tests.
                 cpspec["folder"] = relpath(cpspec["folder"])
-                del cpspec["legacy"]
+                del cpspec[seedless_types[test.index(True)]]
+                ## Fix this latter
                 self.seedless[cpspec["name"]] = LegacyDatabase(**cpspec)
+                ## Fix this latter
             else:
                 for cspec in self.specs["configs"]:
                     name, poscar = cspec["name"], cspec["poscar"]
@@ -543,12 +599,11 @@ class Controller(object):
 
                     if ("configs" in dbspec) and (cspec["name"] not in dbspec["configs"]):
                         continue
-                    dbname = '.'.join((name, dbspec["name"]))
+                    dbname = dbspec["name"]
                     steps = dbspec["steps"]
-                    seq = Repeater(dbname, poscar, self.root, self,
-                                           steps, dbspec.get("niterations"),
-                                           self.specs.get("splits"))
-                    self.seeded[name][dbspec["name"]] = seq
+                    db = Database(dbname, self.root, self,
+                                  steps, self.specs.get("splits"))
+                    self.seeded[name][dbspec["name"]] = db
 
         from os import mkdir
         if not path.isdir(self.plotdir):